Moved include files to proper location; rework CMakeLists.txt
