+++ packages/core/include/is/core/Message.hpp
