+++ packages/core/include/is/core/runtime/StringTemplate.hpp
