--- packages/core/templates/soss-idl-middleware-mix-config.cmake.in
@@ -1,16 +1,0 @@
-# - Config file for the soss-@_ARG_IDLE_TYPE@-@middleware@-mix package that enables
-#   @_ARG_IDLE_TYPE@ messages to be translated into the @middleware@ middleware
-
-cmake_minimum_required(VERSION 3.5.1 FATAL_ERROR)
-
-if(soss-@_ARG_IDL_TYPE@-@middleware@-mix_INCLUDED)
-  return()
-endif()
-set(soss-@_ARG_IDL_TYPE@-@middleware@-mix_INCLUDED TRUE)
-
-set(SOSS_@_ARG_IDL_TYPE@_@middleware@_EXTENSION "${CMAKE_CURRENT_LIST_DIR}/soss-@_ARG_IDL_TYPE@-@middleware@-mix-extension.cmake")
-set(SOSS_@_ARG_IDL_TYPE@_@middleware@_TEMPLATE_DIR "${CMAKE_CURRENT_LIST_DIR}/templates")
-
-foreach(dep @_ARG_DEPENDENCIES@)
-  find_package(${dep} REQUIRED)
-endforeach()
