+++ packages/core/test/unit/search_test.cpp
@@ -15,23 +15,27 @@
  *
  */
 
-#include <Search-impl.hpp>
+#include <is/core/runtime/Search.hpp>
 
 #include <catch2/catch.hpp>
 
 #include <iostream>
 
-TEST_CASE("Use config-file directory", "[search][core]")
+TEST_CASE("Use config-file directory", "[search][is-core]")
 {
-    soss::Search::Implementation::set_config_file_directory(
+    eprosima::is::core::Search::set_config_file_directory(
         SEARCH_TEST__MOCK_CONFIG_DIRECTORY);
 
-    const soss::Search no_config_dir = soss::Search("mock");
+    const eprosima::is::core::Search no_config_dir =
+            eprosima::is::core::Search("mock");
+
     const auto result = no_config_dir.find_file(SEARCH_TEST__MOCK_FILE_NAME);
     REQUIRE(result.empty());
 
-    const soss::Search with_config_dir = soss::Search("mock")
+    const eprosima::is::core::Search with_config_dir =
+            eprosima::is::core::Search("mock")
             .relative_to_config();
+
     REQUIRE(with_config_dir.find_file(SEARCH_TEST__MOCK_FILE_NAME)
             == SEARCH_TEST__MOCK_FILE_PATH);
 }
